pyproject.toml
[tool.poetry]
name = "toire-raspi-py"
version = "0.1.0"
description = ""
authors = ["Rate980 <xv0nnj0@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyaudio = "^0.2.14"
numpy = "^2.0.1"
pyserial = "^3.5"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
